adding tests for write operations
